Add auto_update for kubevirt app.
This commit adds the auto update capability to the kubevirt app.
This change will enable the kubevirt app to be updated when a
new version is made available to the system.

PASS: build-pkgs -p stx-kubevirt-app-helm
PASS: system application-apply kubevirt-app
PASS: In a system where kubevirt is in applied state, verify
      whether auto_update works by placing the new tarball
      in "/usr/local/share/applications/helm/" and monitoring
      if sysinv automaticaly updates kubevirt to a new version.

Story: 2010628
Task: 48741

Change-Id: Ib82b8d55a41ffc496bd065f142e1993bd787de0b
Signed-off-by: David Barbosa Bastos <david.barbosabastos@windriver.com>